master-agent/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
redis>=5.0.0
celery>=5.3.0
prisma>=0.12.0
asyncpg>=0.29.0

# AI Clients
anthropic>=0.64.0
google-genai>=1.0.0
tenacity>=8.2.0
